Modularize click on first child
