--- e2e-tests/utils/uploadFile.js
@@ -1,7 +1,0 @@
-module.exports = {
-  upload: async function (browser, pathToFile) {
-    // Assumes you have already logged in
-    // Navigate to upload URL
-    //
-  },
-};
